refactor: move the list of all section types into sectionTypes/index
